Picking up items done

+++ server_files/server_item.cpp
@@ -10,26 +10,26 @@ uint8_t Item::get_id() {
     return id;
 }
 
-int Item::get_posX() {
+int16_t Item::get_posX() {
     return id;
 }
 
-int Item::get_posY() {
+int16_t Item::get_posY() {
     return id;
 }
 
-int Item::get_amount() {
+int16_t Item::get_amount() {
     return amount;
 }
 
-void Item::set_amount(int item_amount) {
+void Item::set_amount(int16_t item_amount) {
     amount = item_amount;
 }
 
-void Item::set_posX(int pos) {
+void Item::set_posX(int16_t pos) {
     posX = pos;
 }
 
-void Item::set_posY(int pos) {
+void Item::set_posY(int16_t pos) {
     posY = pos;
 }
